LLM-Optimizer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-optimizer"
version = "0.1.0"
description = "MVP architecture for LLM cost optimization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
production = [
	"openai>=1.30.0",
	"redis>=5.0.0",
	"azure-identity>=1.17.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
